requirements.txt
# TODO: Limpar arquivos de dependências
asgiref==3.3.1
attrs==20.3.0
awsebcli==3.19.2
bcrypt==3.2.0
blessed==1.17.12
boto3==1.16.30
botocore==1.19.32
cached-property==1.5.2
cairocffi==1.2.0
CairoSVG==2.5.0
cement==2.8.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.2.5
cryptography==3.3
cssselect2==0.4.1
defusedxml==0.6.0
distro==1.5.0
Django==3.1.4
django-bootstrap-form==3.4
django-crispy-forms==1.10.0
django-discover-runner==1.0
django-extra-fields==3.0.2
django-onesignal==1.0.0
django-request==1.5.6
django-role-permissions==3.1.1
django-storages==1.10.1
django-weasyprint==1.0.2
djangorestframework==3.12.2
docker==4.4.0
docker-compose>=1.25.2
dockerpty==0.4.1
docopt==0.6.2
future>=0.16.0
gunicorn==20.0.4
h11==0.11.0
html5lib==1.1
httpcore==0.12.2
httpx==0.16.1
idna==2.10
jmespath==0.10.0
jsonschema==3.2.0
onesignal-sdk==2.0.0
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyMySQL==1.0.2
PyNaCl==1.4.0
Pyphen==0.10.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
requests>=2.20.1
rfc3986==1.4.0
s3transfer==0.3.3
semantic-version==2.5.0
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
tinycss2==1.1.0
upgrade-requirements==1.7.0
urllib3>=1.25.4
wcwidth>=0.1.7
WeasyPrint==52.2
webencodings==0.5.1
websocket-client==0.57.0
